update docstrings and rename files
