+++ src/modules/quadraticVoting.ts
@@ -1,8 +1,10 @@
+/**
+ * Calculates quadratic votes for each agent based on the square root of their votes.
+ * @param votes A dictionary representing the number of votes for each user.
+ * @returns A tuple containing a dictionary of quadratic votes for each user and the sum of quadratic votes.
+ * @throws {Error} If a vote for a user is undefined.
+ */
 export function quadraticVoting(votes: Record<string, number>): [Record<string, number>, number] {
-  // Calculates quadratic votes for each agent based on the square root of their votes.
-  // :param: (votes): A dictionary representing the number of votes for each user.
-  // :returns: A tuple containing a dictionary of quadratic votes for each user and the sum of quadratic votes.
-
   const quadraticVotesDict: Record<string, number> = {};
 
   for (const userId in votes) {
